fix: 修复编译错误
- 添加 flutter/material.dart import 到 storage_service.dart
- 修复 takeLast 方法不存在问题
- 修复 batch_processing_screen.dart 类型转换错误

+++ lib/screens/batch_processing_screen.dart
@@ -537,7 +537,7 @@ class _BatchProcessingScreenState extends State<BatchProcessingScreen> {
     String? estimatedRemaining;
     if (_processStartTime != null && progressValue > 0 && progressValue < 1) {
       final elapsed = DateTime.now().difference(_processStartTime!);
-      final totalEstimated = elapsed ~/ progressValue;
+      final totalEstimated = elapsed ~/ progressValue.toInt();
       final remaining = totalEstimated - elapsed;
       if (remaining.inSeconds > 0) {
         final remMinutes = remaining.inMinutes;
